refactor(web): おみくじ装置をパターンレジストリ構成に分離 (#200)
共通フレーム(外壁・鈴と鈴の緒・狐・ビン)を omikujiPatterns/frame.js に、
既存装置v3の中間からくりを pattern01Karakuri.js に移設。
buildMachineWorld はパターン未指定なら毎回ランダムに選ぶ
(現時点では1パターンのため挙動は従来と同一)。
検証スクリプトは全パターンを掃引するよう拡張。
